Fix styles and add basic demo.
The demo was using all of Semantic UI’s styles, which isn’t always wanted in an app that wants to use this module by itself.

This fixes a bug that was caused by not including transition styles. The demo was only working because transition.css is included in the semantic-ui-css styles.  The package itself imports all of its dependencies, now, and demo styles are more targeted to individual packages that are needed.

To make this work semantic-ui-css has been moved from devDependencies to dependencies in the package.json.